move auth scaffold
