+++ source/ui/src/main/java/com/clerk/ui/core/appbar/ClerkTopAppbar.kt
@@ -11,10 +11,12 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.PreviewLightDark
 import com.clerk.ui.R
 import com.clerk.ui.core.avatar.OrganizationAvatar
@@ -27,6 +29,7 @@ fun ClerkTopAppBar(
   modifier: Modifier = Modifier,
   hasLogo: Boolean = true,
   hasBackButton: Boolean = true,
+  title: String? = null,
   onBackPressed: () -> Unit,
 ) {
   ClerkMaterialTheme {
@@ -44,6 +47,13 @@ fun ClerkTopAppBar(
       }
 
       Spacer(modifier = Modifier.weight(1f))
+      title?.let {
+        Text(
+          text = it,
+          style = ClerkMaterialTheme.typography.titleLarge.copy(fontWeight = FontWeight.SemiBold),
+          color = ClerkMaterialTheme.colors.foreground,
+        )
+      }
       if (hasLogo) {
         OrganizationAvatar()
       }
@@ -71,6 +81,21 @@ private fun Preview() {
   ClerkMaterialTheme {
     Box(modifier = Modifier.background(color = ClerkMaterialTheme.colors.background)) {
       ClerkTopAppBar(onBackPressed = {}, hasBackButton = false)
+    }
+  }
+}
+
+@PreviewLightDark
+@Composable
+private fun PreviewTitle() {
+  ClerkMaterialTheme {
+    Box(modifier = Modifier.background(color = ClerkMaterialTheme.colors.background)) {
+      ClerkTopAppBar(
+        onBackPressed = {},
+        hasBackButton = true,
+        title = "Add email address",
+        hasLogo = false,
+      )
     }
   }
 }
