+++ source/ui/src/main/java/com/clerk/ui/core/common/scaffold/ClerkThemedAuthScaffold.kt
@@ -1,4 +1,4 @@
-package com.clerk.ui.core.common
+package com.clerk.ui.core.common.scaffold
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
@@ -20,6 +20,10 @@ import com.clerk.ui.core.appbar.ClerkTopAppBar
 import com.clerk.ui.core.button.standard.ClerkButton
 import com.clerk.ui.core.button.standard.ClerkButtonConfiguration
 import com.clerk.ui.core.button.standard.ClerkButtonDefaults
+import com.clerk.ui.core.common.HeaderTextView
+import com.clerk.ui.core.common.HeaderType
+import com.clerk.ui.core.common.SecuredByClerkView
+import com.clerk.ui.core.common.Spacers
 import com.clerk.ui.core.common.dimens.dp18
 import com.clerk.ui.core.common.dimens.dp32
 import com.clerk.ui.core.error.ClerkErrorSnackbar
