+++ source/ui/src/main/java/com/clerk/ui/signin/alternativemethods/SignInFactorAlternativeMethodsView.kt
@@ -22,9 +22,9 @@ import com.clerk.ui.auth.LocalAuthState
 import com.clerk.ui.auth.PreviewAuthStateProvider
 import com.clerk.ui.core.button.social.ClerkSocialRow
 import com.clerk.ui.core.common.AuthStateEffects
-import com.clerk.ui.core.common.ClerkThemedAuthScaffold
 import com.clerk.ui.core.common.Spacers
 import com.clerk.ui.core.common.StrategyKeys
+import com.clerk.ui.core.common.scaffold.ClerkThemedAuthScaffold
 import com.clerk.ui.core.divider.TextDivider
 import com.clerk.ui.signin.password.forgot.AlternativeFactorList
 import com.clerk.ui.util.TextIconHelper
